% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COV_GENOTYPIC}
\alias{COV_GENOTYPIC}
\title{Covariance Type Code: Genotypic variance-covariance}
\format{
An object of class \code{integer} of length 1.
}
\usage{
COV_GENOTYPIC
}
\description{
Covariance Type Code: Genotypic variance-covariance
}
\keyword{internal}
