% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{COV_PHENOTYPIC}
\alias{COV_PHENOTYPIC}
\title{Covariance Type Code: Phenotypic variance-covariance}
\format{
An object of class \code{integer} of length 1.
}
\usage{
COV_PHENOTYPIC
}
\description{
Covariance Type Code: Phenotypic variance-covariance
}
\keyword{internal}
