% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_correction_factor}
\alias{cpp_correction_factor}
\title{Correction Factor Matrix}
\usage{
cpp_correction_factor(total_sums, n_obs)
}
\arguments{
\item{total_sums}{Vector of column sums}

\item{n_obs}{Number of observations}
}
\value{
Symmetric correction factor matrix
}
\description{
Computes the correction factor matrix for ANOVA calculations.
CF(i,j) = (sum_i * sum_j) / n
}
\keyword{internal}
