% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_correction_factor_matrix}
\alias{cpp_correction_factor_matrix}
\title{Compute Correction Factor Matrix}
\usage{
cpp_correction_factor_matrix(data_mat)
}
\arguments{
\item{data_mat}{Numeric matrix (n_obs x n_traits)}
}
\value{
Matrix of correction factors (n_traits x n_traits)
}
\description{
Computes correction factor for all trait pairs.
CF[i,j] = (sum_i * sum_j) / n_obs
}
\keyword{internal}
