% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_grand_means}
\alias{cpp_grand_means}
\title{Compute Grand Means}
\usage{
cpp_grand_means(data_mat)
}
\arguments{
\item{data_mat}{Numeric matrix (n_obs x n_traits)}
}
\value{
Vector of grand means (n_traits)
}
\description{
Computes mean for each trait (column).
}
\keyword{internal}
