% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_grouped_sum_of_products}
\alias{cpp_grouped_sum_of_products}
\title{Grouped Sum of Products}
\usage{
cpp_grouped_sum_of_products(group_sums, group_counts, CF)
}
\arguments{
\item{group_sums}{Matrix of group sums (n_groups x n_traits)}

\item{group_counts}{Vector of group sizes}

\item{CF}{Correction factor matrix}
}
\value{
Symmetric sum of products matrix
}
\description{
Computes the sum of products for grouped data.
GSP(i,j) = sum_g [(sum_i_g * sum_j_g) / n_g] - CF(i,j)
}
\keyword{internal}
