% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_mean_squares}
\alias{cpp_mean_squares}
\title{Mean Squares from Sum of Products}
\usage{
cpp_mean_squares(sum_of_products, df)
}
\arguments{
\item{sum_of_products}{Sum of products matrix}

\item{df}{Degrees of freedom}
}
\value{
Mean squares matrix
}
\description{
Divides sum of products matrix by degrees of freedom.
MS = SP / df
}
\keyword{internal}
