% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_quadratic_form}
\alias{cpp_quadratic_form}
\title{Quadratic Form: x' A y}
\usage{
cpp_quadratic_form(x, A, y)
}
\arguments{
\item{x}{First vector}

\item{A}{Matrix}

\item{y}{Second vector}
}
\value{
Scalar result of x' A y
}
\description{
Computes the quadratic form x' A y efficiently.
Equivalent to t(x) %*% A %*% y in R but optimized.
}
\keyword{internal}
