% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_quadratic_form_sym}
\alias{cpp_quadratic_form_sym}
\title{Symmetric Quadratic Form: x' A x}
\usage{
cpp_quadratic_form_sym(x, A)
}
\arguments{
\item{x}{Vector}

\item{A}{Symmetric matrix}
}
\value{
Scalar result of x' A x
}
\description{
Computes the symmetric quadratic form x' A x efficiently.
Equivalent to t(x) %*% A %*% x in R but optimized.
}
\keyword{internal}
