% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_total_sum_of_products}
\alias{cpp_total_sum_of_products}
\title{Total Sum of Products}
\usage{
cpp_total_sum_of_products(data_mat, CF)
}
\arguments{
\item{data_mat}{Data matrix (n_obs x n_traits)}

\item{CF}{Correction factor matrix}
}
\value{
Symmetric sum of products matrix
}
\description{
Computes the total sum of products matrix corrected for the mean.
TSP(i,j) = sum(x_i * x_j) - CF(i,j)
}
\keyword{internal}
