% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_trait_minmax}
\alias{cpp_trait_minmax}
\title{Compute Trait-wise Min/Max}
\usage{
cpp_trait_minmax(data_mat)
}
\arguments{
\item{data_mat}{Numeric matrix (n_obs x n_traits)}
}
\value{
List with 'min' and 'max' vectors
}
\description{
Computes minimum and maximum for each trait.
}
\keyword{internal}
