% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESIGN_CODES}
\alias{DESIGN_CODES}
\title{Vector of all valid design codes}
\format{
An object of class \code{integer} of length 3.
}
\usage{
DESIGN_CODES
}
\description{
Vector of all valid design codes
}
\keyword{internal}
