% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESIGN_LSD}
\alias{DESIGN_LSD}
\title{Design Type Code: Latin Square Design (LSD)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DESIGN_LSD
}
\description{
Design Type Code: Latin Square Design (LSD)
}
\keyword{internal}
