% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESIGN_NAMES}
\alias{DESIGN_NAMES}
\title{Named vector mapping design names to codes}
\format{
An object of class \code{integer} of length 3.
}
\usage{
DESIGN_NAMES
}
\description{
Named vector mapping design names to codes
}
\keyword{internal}
