% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESIGN_RCBD}
\alias{DESIGN_RCBD}
\title{Design Type Code: Randomized Complete Block Design (RCBD)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DESIGN_RCBD
}
\description{
Design Type Code: Randomized Complete Block Design (RCBD)
}
\keyword{internal}
