% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{DESIGN_SPD}
\alias{DESIGN_SPD}
\title{Design Type Code: Split Plot Design (SPD)}
\format{
An object of class \code{integer} of length 1.
}
\usage{
DESIGN_SPD
}
\description{
Design Type Code: Split Plot Design (SPD)
}
\keyword{internal}
