% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_phenotypic_indices.R
\name{.cochran_adjustment}
\alias{.cochran_adjustment}
\title{Compute Cochran/Cunningham covariance adjustment}
\usage{
.cochran_adjustment(cov_xy, cov_xw, cov_yw, var_w, k1, tau)
}
\description{
Compute Cochran/Cunningham covariance adjustment
}
\keyword{internal}
