% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_genomic_indices.R
\name{.cochran_adjustment_genomic}
\alias{.cochran_adjustment_genomic}
\title{Compute Cochran/Cunningham covariance adjustment for genomic matrices}
\usage{
.cochran_adjustment_genomic(Gamma, Gamma1, beta1, A, k1, tau)
}
\description{
Compute Cochran/Cunningham covariance adjustment for genomic matrices
}
\keyword{internal}
