% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_genomic_indices.R
\name{.combined_index_metrics}
\alias{.combined_index_metrics}
\title{Compute combined index metrics (for CRLGSI/CPPG-LGSI)}
\usage{
.combined_index_metrics(
  b,
  T_C,
  Psi_C,
  w = NULL,
  k_I = 2.063,
  L_I = 1,
  GAY = NULL
)
}
\description{
Compute combined index metrics (for CRLGSI/CPPG-LGSI)
}
\keyword{internal}
