% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{.compute_genetic_values}
\alias{.compute_genetic_values}
\title{Compute Genetic Values for Population}
\usage{
.compute_genetic_values(haplotype1, haplotype2, qtl_effects)
}
\arguments{
\item{haplotype1}{First haplotype matrix (n_individuals x n_loci)}

\item{haplotype2}{Second haplotype matrix (n_individuals x n_loci)}

\item{qtl_effects}{QTL effect matrix (n_loci x n_traits)}
}
\value{
Matrix of genetic values (n_individuals x n_traits)
}
\description{
Calculates breeding values (genetic values) for all individuals in the
population based on their diploid genotypes and QTL effects.
}
\keyword{internal}
