% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{.esim_solve_sym_multi}
\alias{.esim_solve_sym_multi}
\title{Solve symmetric linear system for multiple right-hand sides}
\usage{
.esim_solve_sym_multi(A, B)
}
\description{
Solve symmetric linear system for multiple right-hand sides
}
\keyword{internal}
