% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{.generate_offspring}
\alias{.generate_offspring}
\title{Generate Offspring Population Through Recombination}
\usage{
.generate_offspring(selected_parents, n_offspring)
}
\arguments{
\item{selected_parents}{List containing haplotypes and metadata of selected individuals}

\item{n_offspring}{Number of offspring to generate}
}
\value{
List with offspring haplotypes
}
\description{
Creates the next generation by randomly mating selected parents and
simulating meiotic recombination.
}
\keyword{internal}
