% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_genomic_indices.R
\name{.genomic_index_correlation}
\alias{.genomic_index_correlation}
\title{Compute genomic index correlation}
\usage{
.genomic_index_correlation(beta1, beta2, Gamma1, Gamma, A)
}
\description{
Compute genomic index correlation
}
\keyword{internal}
