% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_genomic_indices.R
\name{.genomic_index_metrics}
\alias{.genomic_index_metrics}
\title{Compute genomic index metrics using C++ primitives}
\usage{
.genomic_index_metrics(
  b,
  Gamma,
  w = NULL,
  k_I = 2.063,
  L_G = 1,
  GAY = NULL,
  gmat = NULL
)
}
\description{
Compute genomic index metrics using C++ primitives
}
\keyword{internal}
