% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_genomic_indices.R
\name{.genomic_stage_metrics}
\alias{.genomic_stage_metrics}
\title{Compute genomic stage metrics for multistage indices}
\usage{
.genomic_stage_metrics(beta, Gamma, A, w, k)
}
\description{
Compute genomic stage metrics for multistage indices
}
\keyword{internal}
