% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_indices.R
\name{.index_metrics}
\alias{.index_metrics}
\title{Compute selection index metrics using C++ primitives}
\usage{
.index_metrics(b, P, G, w = NULL, const_factor = 2.063, GAY = NULL)
}
\description{
Compute selection index metrics using C++ primitives
}
\keyword{internal}
