% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{.recombine_haplotypes}
\alias{.recombine_haplotypes}
\title{Recombine Two Haplotypes}
\usage{
.recombine_haplotypes(haplotype1, haplotype2, recombination_fractions)
}
\arguments{
\item{haplotype1}{First parental haplotype (vector of length n_loci)}

\item{haplotype2}{Second parental haplotype (vector of length n_loci)}

\item{recombination_fractions}{Vector of recombination fractions between
adjacent loci (length n_loci - 1)}
}
\value{
Recombinant haplotype (vector of length n_loci)
}
\description{
Simulates meiotic recombination between two parental haplotypes using
the recombination fractions derived from Haldane's mapping function.
}
\keyword{internal}
