% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_indices.R
\name{.solve_sym_multi}
\alias{.solve_sym_multi}
\title{Solve symmetric system for multiple right-hand sides}
\usage{
.solve_sym_multi(A, B)
}
\description{
Solve symmetric system for multiple right-hand sides
}
\keyword{internal}
