% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_phenotypic_indices.R
\name{.stage_metrics}
\alias{.stage_metrics}
\title{Compute stage metrics for multistage indices}
\usage{
.stage_metrics(b, P, G, w, k)
}
\description{
Compute stage metrics for multistage indices
}
\keyword{internal}
