% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_genomic_indices.R
\name{.young_intensities}
\alias{.young_intensities}
\title{Young's method for selection intensities (same as phenotypic)}
\usage{
.young_intensities(p, rho_12)
}
\description{
Young's method for selection intensities (same as phenotypic)
}
\keyword{internal}
