% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multistage_phenotypic_indices.R
\name{.young_selection_intensities}
\alias{.young_selection_intensities}
\title{Young's method for selection intensities}
\usage{
.young_selection_intensities(p, rho_12)
}
\description{
Young's method for selection intensities
}
\keyword{internal}
