% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.advance.R
\name{gen.advance}
\alias{gen.advance}
\title{Genetic Advance for PRE}
\usage{
gen.advance(phen_mat, gen_mat, weight_mat)
}
\arguments{
\item{phen_mat}{phenotypic matrix value of desired characters}

\item{gen_mat}{genotypic matrix value of desired characters}

\item{weight_mat}{weight matrix value of desired characters}
}
\value{
Genetic advance of character or character combinations
}
\description{
Genetic Advance for PRE
}
\examples{
gmat<- gen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat<- phen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
gen.advance(phen_mat = pmat[1,1], gen_mat = gmat[1,1], weight_mat = weight[1,2])
}
