% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.varcov.R
\name{gen.varcov}
\alias{gen.varcov}
\title{Genotypic Variance-Covariance Analysis}
\usage{
gen.varcov(data, genotypes, replication)
}
\arguments{
\item{data}{traits to be analyzed}

\item{genotypes}{vector containing genotypes/treatments}

\item{replication}{vector containing replication}
}
\value{
A Genotypic Variance-Covariance Matrix
}
\description{
Genotypic Variance-Covariance Analysis
}
\examples{
gen.varcov(data=seldata[,3:9], genotypes=seldata$treat,replication=seldata$rep)
}
