% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpsi.R
\name{lpsi}
\alias{lpsi}
\title{Construction of selection indices (filtered/combinatorial)}
\usage{
lpsi(ncomb, pmat, gmat, wmat, wcol = 1, GAY, excluding_trait = NULL)
}
\arguments{
\item{ncomb}{Number of Characters/Traits group}

\item{pmat}{Phenotypic Variance-Covariance Matrix}

\item{gmat}{Genotypic Variance-Covariance Matrix}

\item{wmat}{Weight Matrix}

\item{wcol}{Weight column number incase more than one weights, by default its 1}

\item{GAY}{Genetic Advance of comparative Character/Trait i.e. Yield (Optional argument)}

\item{excluding_trait}{Optional. Traits to exclude from combinations. Can be:
(1) numeric vector of trait indices (e.g., c(1, 3)),
(2) character vector of trait names (e.g., c("sypp", "dtf")),
(3) data frame/matrix columns with trait data (trait names extracted from column names).
When specified, only combinations that do NOT contain any of these traits are returned.}
}
\value{
Data frame of all possible selection indices with metrics (GA, PRE, Delta_G, rHI, hI2)
}
\description{
Build all possible selection indices from trait combinations, with optional
exclusion of specific traits (by index, name, or data columns).
}
\examples{
gmat<- gen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat<- phen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
wmat<- weight_mat(weight)
lpsi(ncomb = 1, pmat = pmat, gmat = gmat, wmat = wmat, wcol = 1, GAY = 1.075)
}
