% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{plot.selection_simulation}
\alias{plot.selection_simulation}
\title{Plot Method for Selection Simulation Results}
\usage{
\method{plot}{selection_simulation}(x, trait_index = 1, type = c("mean", "gain"), ...)
}
\arguments{
\item{x}{Object of class \code{selection_simulation}}

\item{trait_index}{Which trait to plot (default: 1)}

\item{type}{Type of plot: "mean" for mean genetic value, "gain" for genetic gain per cycle}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot Method for Selection Simulation Results
}
\keyword{internal}
