% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{ppg_lpsi}
\alias{ppg_lpsi}
\title{Predetermined Proportional Gains (PPG-LPSI)}
\usage{
ppg_lpsi(pmat, gmat, k, wmat = NULL, wcol = 1, GAY)
}
\arguments{
\item{pmat}{Phenotypic variance-covariance matrix}

\item{gmat}{Genotypic variance-covariance matrix}

\item{k}{Vector of desired proportional gains}

\item{wmat}{Optional weight matrix for GA/PRE calculation}

\item{wcol}{Weight column number (default: 1)}

\item{GAY}{Genetic advance of comparative trait (optional)}
}
\value{
List with summary data frame, coefficient vector, Delta_G vector, and phi
}
\description{
Predetermined Proportional Gains (PPG-LPSI)
}
\examples{
gmat <- gen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat <- phen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
k <- rep(1, ncol(pmat))
ppg_lpsi(pmat, gmat, k)
}
