% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{print.base_index}
\alias{print.base_index}
\title{Print method for Base Index}
\usage{
\method{print}{base_index}(x, ...)
}
\arguments{
\item{x}{Object of class 'base_index'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for Base Index
}
