% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{print.dg_lpsi}
\alias{print.dg_lpsi}
\title{Print method for Desired Gains Index}
\usage{
\method{print}{dg_lpsi}(x, ...)
}
\arguments{
\item{x}{Object of class 'dg_lpsi'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for Desired Gains Index
}
