% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{print.esim}
\alias{print.esim}
\title{Print method for ESIM}
\usage{
\method{print}{esim}(x, ...)
}
\arguments{
\item{x}{Object of class 'esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for ESIM
}
