% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{print.gesim}
\alias{print.gesim}
\title{Print method for GESIM}
\usage{
\method{print}{gesim}(x, ...)
}
\arguments{
\item{x}{Object of class 'gesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for GESIM
}
