% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{print.gw_esim}
\alias{print.gw_esim}
\title{Print method for GW-ESIM}
\usage{
\method{print}{gw_esim}(x, ...)
}
\arguments{
\item{x}{Object of class 'gw_esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for GW-ESIM
}
