% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{print.mesim}
\alias{print.mesim}
\title{Print method for MESIM}
\usage{
\method{print}{mesim}(x, ...)
}
\arguments{
\item{x}{Object of class 'mesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for MESIM
}
