% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{print.ppg_esim}
\alias{print.ppg_esim}
\title{Print method for PPG-ESIM}
\usage{
\method{print}{ppg_esim}(x, ...)
}
\arguments{
\item{x}{Object of class 'ppg_esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for PPG-ESIM
}
