% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{print.ppg_gesim}
\alias{print.ppg_gesim}
\title{Print method for PPG-GESIM}
\usage{
\method{print}{ppg_gesim}(x, ...)
}
\arguments{
\item{x}{Object of class 'ppg_gesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for PPG-GESIM
}
