% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{print.resim}
\alias{print.resim}
\title{Print method for RESIM}
\usage{
\method{print}{resim}(x, ...)
}
\arguments{
\item{x}{Object of class 'resim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for RESIM
}
