% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{print.rgesim}
\alias{print.rgesim}
\title{Print method for RGESIM}
\usage{
\method{print}{rgesim}(x, ...)
}
\arguments{
\item{x}{Object of class 'rgesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for RGESIM
}
