% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_indices.R
\name{print.smith_hazel}
\alias{print.smith_hazel}
\title{Print method for Smith-Hazel Index}
\usage{
\method{print}{smith_hazel}(x, ...)
}
\arguments{
\item{x}{Object of class 'smith_hazel'}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for Smith-Hazel Index
}
