% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcomb.indices.R
\name{rcomb.indices}
\alias{rcomb.indices}
\title{Remove trait or trait combination from possible trait combinations of possible
Trait combinations}
\usage{
rcomb.indices(ncomb, i, pmat, gmat, wmat, wcol = 1, GAY)
}
\arguments{
\item{ncomb}{Number of character combination}

\item{i}{remove trait or trait combination}

\item{pmat}{Phenotypic Variance Covariance Matrix}

\item{gmat}{Genotypic Variane Covariance Matrix}

\item{wmat}{Weight Matrix}

\item{wcol}{Respective weight column number of Weight Matrix}

\item{GAY}{Genetic Advance/Genetic Gain of base selection index}
}
\value{
Data frame of possible selection indices with per cent relative efficiency and ranking
}
\description{
Remove trait or trait combination from possible trait combinations of possible
Trait combinations
}
\examples{
rcomb.indices(ncomb = 2, i = 1, pmat = pmat, gmat = gmat, wmat = wmat, wcol = 1)
}
