% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcomb_indices.R
\name{rcomb_indices}
\alias{rcomb_indices}
\title{Remove trait or trait combination from possible trait combinations of possible
Trait combinations}
\usage{
rcomb_indices(ncomb, i, pmat, gmat, wmat, wcol = 1, GAY)
}
\arguments{
\item{ncomb}{Number of character combination}

\item{i}{remove trait or trait combination}

\item{pmat}{Phenotypic Variance Covariance Matrix}

\item{gmat}{Genotypic Variance Covariance Matrix}

\item{wmat}{Weight Matrix}

\item{wcol}{Respective weight column number of Weight Matrix}

\item{GAY}{Genetic Advance/Genetic Gain of base selection index}
}
\value{
Data frame of possible selection indices with per cent relative efficiency and ranking
}
\description{
Remove trait or trait combination from possible trait combinations of possible
Trait combinations
}
\examples{
gmat<- gen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat<- phen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
rcomb_indices(ncomb = 2, i = 1, pmat = pmat, gmat = gmat, wmat = weight[,2:3], wcol = 1)
}
