% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{rlpsi}
\alias{rlpsi}
\title{Restricted Linear Phenotypic Selection Index (RLPSI)}
\usage{
rlpsi(pmat, gmat, wmat, wcol = 1, C, GAY)
}
\arguments{
\item{pmat}{Phenotypic variance-covariance matrix}

\item{gmat}{Genotypic variance-covariance matrix}

\item{wmat}{Weight matrix}

\item{wcol}{Weight column number (default: 1)}

\item{C}{Constraint matrix (n_traits x n_constraints). Each column is a restriction.}

\item{GAY}{Genetic advance of comparative trait (optional)}
}
\value{
List with summary data frame, coefficient vector, and Delta_G vector
}
\description{
Restricted Linear Phenotypic Selection Index (RLPSI)
}
\examples{
gmat <- gen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat <- phen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
wmat <- weight_mat(weight)
C <- diag(ncol(pmat))[, 1, drop = FALSE]
rlpsi(pmat, gmat, wmat, wcol = 1, C = C)
}
