% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.index.R
\name{sel.index}
\alias{sel.index}
\title{Construction of Selection Index}
\usage{
sel.index(ID, phen_mat, gen_mat, weight_mat, GAY)
}
\arguments{
\item{ID}{Name of Selection Index}

\item{phen_mat}{Phenotypic Variance-Covariance Matrix}

\item{gen_mat}{Genotypic Variance-Covariance Matrix}

\item{weight_mat}{Weight Matrix}

\item{GAY}{Genetic Advance of Yield}
}
\value{
List of Selection Index Components
}
\description{
Construction of Selection Index
}
\examples{
gmat<- gen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat<- phen.varcov(seldata[,3:9], seldata[,2], seldata[,1])
sel.index(ID = 1, phen_mat = pmat[1,1], gen_mat = gmat[1,1], weight_mat = weight[1,2])
}
