% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel.score.rank.R
\name{sel.score.rank}
\alias{sel.score.rank}
\title{Analysis of selection score and Genotype/Progeny ranking based on selection score}
\usage{
sel.score.rank(data, bmat, genotype)
}
\arguments{
\item{data}{traits to be analyzed}

\item{bmat}{Selection/discriminant coefficient matrix}

\item{genotype}{Vector containing Genotypes/Treatments}
}
\value{
A data frame containing the genotypes with selection score
}
\description{
Analysis of selection score and Genotype/Progeny ranking based on selection score
}
\examples{
sel.score.rank(data = seldata[,3], bmat = 0.6455, genotype = seldata[,2])
}
