% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seldata.R
\docType{data}
\name{seldata}
\alias{seldata}
\title{Selection Index DataSet}
\format{
A data frame with 144 rows and 9 columns
}
\usage{
data(seldata)
}
\description{
A dataset containing the data of three replications and 48 progenies
with 7 different traits.
}
\details{
\itemize{
\item rep. Replications
\item treat. Treatments/Genotypes
\item sypp. Seed Yield per Plant
\item dtf. Days to 50% Flowering
\item rpp. Racemes per Plant
\item ppr. Pods per Raceme
\item ppp. Pods per Plant
\item spp. Seeds per Pod
\item pw. Pods Weight
}
}
