% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.index-package.R
\docType{package}
\name{selection.index-package}
\alias{selection.index}
\alias{selection.index-package}
\title{selection.index: Analysis of Selection Index in Plant Breeding}
\description{
Provides tools for the simultaneous improvement of multiple traits in plant breeding. Building upon the classical selection index (Smith 1937 \doi{10.1111/j.1469-1809.1936.tb02143.x}) and modern quantitative genetics (Kang 2020 \doi{10.1007/978-3-319-91223-3}), this package calculates classical phenotypic, genomic, marker-assisted, restricted/constrained, and eigen selection indices. It also incorporates multi-stage selection evaluation and stochastic simulations to estimate genetic advance based on economic weights, heritability, and genetic correlations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zankrut20/selection.index}
  \item \url{https://zankrut20.github.io/selection.index/}
  \item Report bugs at \url{https://github.com/zankrut20/selection.index/issues}
}

}
\author{
\strong{Maintainer}: Zankrut Goyani \email{zankrut20@gmail.com} [copyright holder]

}
\keyword{internal}
