% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{summary.base_index}
\alias{summary.base_index}
\title{Summary method for Base Index}
\usage{
\method{summary}{base_index}(object, ...)
}
\arguments{
\item{object}{Object of class 'base_index'}

\item{...}{Additional arguments passed to print}
}
\description{
Summary method for Base Index
}
