% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{summary.dg_lpsi}
\alias{summary.dg_lpsi}
\title{Summary method for Desired Gains Index}
\usage{
\method{summary}{dg_lpsi}(object, ...)
}
\arguments{
\item{object}{Object of class 'dg_lpsi'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for Desired Gains Index
}
