% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{summary.esim}
\alias{summary.esim}
\title{Summary method for ESIM}
\usage{
\method{summary}{esim}(object, ...)
}
\arguments{
\item{object}{Object of class 'esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for ESIM
}
