% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{summary.gesim}
\alias{summary.gesim}
\title{Summary method for GESIM}
\usage{
\method{summary}{gesim}(object, ...)
}
\arguments{
\item{object}{Object of class 'gesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for GESIM
}
