% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{summary.gw_esim}
\alias{summary.gw_esim}
\title{Summary method for GW-ESIM}
\usage{
\method{summary}{gw_esim}(object, ...)
}
\arguments{
\item{object}{Object of class 'gw_esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for GW-ESIM
}
