% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{summary.mesim}
\alias{summary.mesim}
\title{Summary method for MESIM}
\usage{
\method{summary}{mesim}(object, ...)
}
\arguments{
\item{object}{Object of class 'mesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for MESIM
}
