% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{summary.ppg_esim}
\alias{summary.ppg_esim}
\title{Summary method for PPG-ESIM}
\usage{
\method{summary}{ppg_esim}(object, ...)
}
\arguments{
\item{object}{Object of class 'ppg_esim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for PPG-ESIM
}
