% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{summary.ppg_gesim}
\alias{summary.ppg_gesim}
\title{Summary method for PPG-GESIM}
\usage{
\method{summary}{ppg_gesim}(object, ...)
}
\arguments{
\item{object}{Object of class 'ppg_gesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for PPG-GESIM
}
