% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_indices.R
\name{summary.resim}
\alias{summary.resim}
\title{Summary method for RESIM}
\usage{
\method{summary}{resim}(object, ...)
}
\arguments{
\item{object}{Object of class 'resim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for RESIM
}
