% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_eigen_indices.R
\name{summary.rgesim}
\alias{summary.rgesim}
\title{Summary method for RGESIM}
\usage{
\method{summary}{rgesim}(object, ...)
}
\arguments{
\item{object}{Object of class 'rgesim'}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for RGESIM
}
