% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_simulation.R
\name{summary.selection_simulation}
\alias{summary.selection_simulation}
\title{Summary Method for Selection Simulation Results}
\usage{
\method{summary}{selection_simulation}(object, ...)
}
\arguments{
\item{object}{Object of class \code{selection_simulation}}

\item{...}{Additional arguments (not used)}
}
\description{
Summary Method for Selection Simulation Results
}
