% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypic_indices.R
\name{summary.smith_hazel}
\alias{summary.smith_hazel}
\title{Summary method for Smith-Hazel Index}
\usage{
\method{summary}{smith_hazel}(object, ...)
}
\arguments{
\item{object}{Object of class 'smith_hazel'}

\item{...}{Additional arguments passed to print}
}
\description{
Summary method for Smith-Hazel Index
}
