% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_CONV}
\alias{TOL_CONV}
\title{Tolerance for convergence in iterative algorithms
Default convergence criterion for missing value imputation}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_CONV
}
\description{
Tolerance for convergence in iterative algorithms
Default convergence criterion for missing value imputation
}
\keyword{internal}
