% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_EQUAL}
\alias{TOL_EQUAL}
\title{Tolerance for general floating point comparisons}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_EQUAL
}
\description{
Tolerance for general floating point comparisons
}
\keyword{internal}
