% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_PSD}
\alias{TOL_PSD}
\title{Tolerance for positive semi-definite checks
Eigenvalues greater than -TOL_PSD are considered non-negative}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_PSD
}
\description{
Tolerance for positive semi-definite checks
Eigenvalues greater than -TOL_PSD are considered non-negative
}
\keyword{internal}
