% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_SYM}
\alias{TOL_SYM}
\title{Tolerance for symmetry checks (e.g., isSymmetric)
Used when checking if matrices are symmetric within numerical precision}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_SYM
}
\description{
Tolerance for symmetry checks (e.g., isSymmetric)
Used when checking if matrices are symmetric within numerical precision
}
\keyword{internal}
