% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{TOL_ZERO}
\alias{TOL_ZERO}
\title{Tolerance for zero checks (near-zero values)
Values smaller than this in absolute value are considered zero}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
TOL_ZERO
}
\description{
Tolerance for zero checks (near-zero values)
Values smaller than this in absolute value are considered zero
}
\keyword{internal}
