% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.mat.R
\name{weight.mat}
\alias{weight.mat}
\title{Convert dataframe to matrix}
\usage{
weight.mat(data)
}
\arguments{
\item{data}{dataframe of weight}
}
\value{
A matrix
}
\description{
Convert dataframe to matrix
}
\examples{
weight.mat(data = seldata)
}
