% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight}
\alias{weight}
\title{Weight dataset}
\format{
A data frame with 7 rows and 3 columns
}
\usage{
data(weight)
}
\description{
A dataset containing the data of 2 different weights namely euqal weight and
broad sense heritability
}
\details{
\itemize{
\item EW. Equal Weight
\item h2. Broad Sense Heritability
}
}
