% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained_indices.R
\name{dg_lpsi}
\alias{dg_lpsi}
\title{Desired Gains Index (DG-LPSI)}
\usage{
dg_lpsi(pmat, gmat, d, wmat = NULL, wcol = 1, GAY)
}
\arguments{
\item{pmat}{Phenotypic variance-covariance matrix}

\item{gmat}{Genotypic variance-covariance matrix}

\item{d}{Vector of desired gains}

\item{wmat}{Optional weight matrix for GA/PRE calculation}

\item{wcol}{Weight column number (default: 1)}

\item{GAY}{Genetic advance of comparative trait (optional)}
}
\value{
List with summary data frame, coefficient vector, and Delta_G vector
}
\description{
Desired Gains Index (DG-LPSI)
}
\examples{
gmat <- gen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
pmat <- phen_varcov(seldata[,3:9], seldata[,2], seldata[,1])
d <- rep(1, ncol(pmat))
dg_lpsi(pmat, gmat, d)
}
