% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanPerformance.R
\name{meanPerformance}
\alias{meanPerformance}
\title{Mean performance of phenotypic data}
\usage{
meanPerformance(data, genotypes, replications)
}
\arguments{
\item{data}{data for analysis}

\item{genotypes}{genotypes vector}

\item{replications}{replication vector}
}
\value{
Dataframe of mean performance analysis
}
\description{
Mean performance of phenotypic data
}
\examples{
meanPerformance(data = seldata[, 3:9], genotypes = seldata[, 2], replications = seldata[, 1])


}
