% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phen.varcov.R
\name{phen.varcov}
\alias{phen.varcov}
\title{Phenotypic Variance-Covariance Analysis}
\usage{
phen.varcov(data, genotypes, replication)
}
\arguments{
\item{data}{traits to be analyzed}

\item{genotypes}{vector containing genotypes/treatments}

\item{replication}{vector containing replication}
}
\value{
A Phenotypic Variance-Covariance Matrix
}
\description{
Phenotypic Variance-Covariance Analysis
}
\examples{
phen.varcov(data=seldata[,3:9], genotypes=seldata$treat,replication=seldata$rep)
}
